Decode-Viewer: Simplify structure and pass components to storybook
