+++ examples/decode-viewer/src/stories/ArgStructure/types.ts
@@ -1,0 +1,3 @@
+export type KeyValueType = {
+  [key: string]: object
+}
